requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
pydantic
pinecone
google-genai
sentence-transformers
transformers
torch
pdfplumber
python-docx
beautifulsoup4
requests
nltk
rouge-score
scikit-learn
python-dotenv
tiktoken